Work a bit more on the animation delay

+++ UnityClient/Assets/Scripts/Heimdallr/Core/GameEntity/Mesh/MeshGameEntity.cs
@@ -85,7 +85,7 @@ namespace Heimdallr.Core.Game {
             throw new NotImplementedException();
         }
 
-        public override void SetAction(EntityActionRequest actionRequestAction, bool isSource, float delay = 0f) {
+        public override void SetAction(EntityActionRequest actionRequestAction, bool isSource, long delay = 0) {
             throw new NotImplementedException();
         }
